finishing refactoring of the actions, adding a pause after zoom for legacyActions
